added prior over criterion orientations to model description
One of the central motivations for us to change to a model including the
criterion was that we wanted to model larger uncertainty for oblique criterion
orientations. This was still missing. So I now included a prior over criteria
that favours cardinal orientations by a parametrically modulated amount. Larger
differences in the prior probabilities (or densities) between cardinal and
oblique orientations will propagate through to correspondingly larger posterior
uncertainty over oblique orientations.