checking behavioural differences across direction and criterion values
This is also an older change. This analysis actually triggered the development
of the new difference model. The new figures show that there is a clear effect
in the behavioural data of the criterion: decisions with criterion with
cardinal orientation appear to be easier than with oblique orientations. There
are also effects in the data suggesting that performance differs across motion
directions, but it might be that these, or at least the largest ones, can also
be explained by the effect of the cardinal/oblique criteria.