first tests of inference with new difference model
Inference with the model runs smoothly, but slow. Depending on the number of
free parameters, it can take 3.5 or 4.5 hours to complete for a single subject.

I tried to simultaneously fit cpsqrtkappa, critstd, dnoisestd and dirstd
(201809171452). The posterior over parameters looked sensible, but the RT
distributions did not fit well. Especially, they did not exhibit sensible
differences between easy and hard trials.

Fixing cpsqrtkappa to 2.2 (201809172008) sets the prior probability for
cardinal criteria roughly twice as high as that for oblique criteria. With
fixed cpsqrtkappa the results looked very similar to those which I got before
implementing cpsqrtkappa (corresponding to cpsqrtkappa=0, results in
201809141814). I did not seriously compare these results, though. So this is
only a first intuition.

Unfortunately, I have no clue what all this means. I first need to develop a
better understanding of the model parameters and how they potentially interact.
The only thing that appears to be certain at this stage is that inference
doesn't work as nicely as I want. Also, I still see with the new difference
model that the difference between easy and hard accuracy is overestimated.