new plotting function in helpers: plot_differences
This outsources plotting of the accuracy and median RT difference between two
conditions so that it can be reused in the analysis script. The function also
handles plotting of two results next to each other so they can directly be
compared.

I now also plot the differences for cardinal vs. oblique conditions instead of
only for easy vs. hard in infer_with_snl.py.