several tests of inference with the diff model for subject 19
I have tested inference with the diff model with different combinations of
priors and fixed parameters. Here are the results and how they differ:

201810071739 - cpsqrtkappa prior 0, 1; critstd=0.7
201810080019 - cpsqrtkappa prior 1, 1; critstd=0.7
201810082106 - cpsqrtkappa=0; critstd=0.7
201810090117 - cpsqrtkappa=0; critstd=1

Otherwise parameter settings were equal across those runs.

The commit now includes code for comparing results based on their posterior
predictive differences in accuracy and median RT, but I didn't have that when
generating those results. That's why I thought that the diff model better
captures the accuracy difference between easy and hard models than the base
model, when talking to Dragan last. The issue was that the base model tended to
overestimate the accracy difference when it tried to get the median RT
difference right. With the diff model this didn't appear to occur for subject
19. It now turns out that the same holds already true for the base model for
subject 19. So subject 19 appears to not be the right subject to test this on.

Because cpsqrtkappa appears to be hard to estimate from the data (the
overall RT distributions don't appear to be fit very well) and it doesn't even
contribute much to fitting the accuracy and median RT difference between oblique
and cardinal criteria, we decided to fix cpsqrtkappa to 0 for the last two
tests. The idea was that having uncertainty over the criterion as set by critstd
already leads to a better fit of the data for the easy vs. hard differences.
Whether this is true, is hard to say based on the current results. For critstd=1
the easy/hard difference in accuracy is actually mostly below the data value,
but the posterior predictions have quite large variance exceeding that of the
base model posterior predictions. For critstd=0.7 the easy/hard accuracy
difference actually exceeds that of the base model value, but has lower
variance. I currently don't know whether the overall fit of the RT distributions
was reasonable in both cases.

Proceeding from here probably includes selecting a subject which exhibits a
strong overestimate of the accuracy difference.

One thing
that appears to be true about the diff model posterior is that it has larger
tails in the posterior predictive differences, i.e., there's more variance in
the posterior predictions.