new function for estimating goodness of fit measures from posterior
This compares the observed summary statistics to the ones generated from
the posterior samples during the simulation step of inference. The
comparison measure (deviance) is provided by the user and should be some
kind of error function. Different deviances for aggregation of different
elements of the summary statistics and for aggregation across posterior
samples can be chosen.