comparing fits of HDDM and base model
We wanted to see how the base model compares to the HDDM, i.e., whether it could
fit the data better than the HDDM. This appears to be the case. Given that the
HDDM uses different drift rates for the easy and hard conditions, it is quite
remarkable that the base model achieves a slightly better fit as measured by
estimate_posterior_fit, i.e., from predictive samples from the posterior. So
the computed deviances just measure the fit to the data, not real predictive
performance, and it could be that snl doesn't estimate the correct posterior. If
snl estimates a too narrow posterior and HDDM doesn't, the computed deviances
should also favor the base model over HDDM, but I cannot tell whether this is
the case.

As I saved the HDDM result in a different format than the snl results, I had to
adapt estimate_posterior_fit correspondingly. Also, HDDM results were produced
with python3. So they cannot be read in python2 (incompatible pickle protocol).