fix: generate correct responses with lik-Trials
I before just set the correct response for the passed trials when Trials
contains likelihoods to NaN, because there was no reliable information
to determine which response was correct in these trials. As I'm using the
model's stored correct responses in a later default analysis, I decided
to make this work by passing the true motion directions to the model,
when Trials contains likelihoods.