first test of inference with EEG data as likelihoods
Arbitrarily chose subject 18, because I thought I previously used this
subject for tests, but anyway don't remember. Inference now ran through
successfully and appears to do what it should.

There was significant
posterior correlation between noisestd and intstd, but this was mostly in
the tail of the corresponding distributions, i.e., the marginals had a
visible peak at a lower value so I ignore this for now.