removing noisestd when using likelihoods as Trials input
It was completely nonsensical to simultaneously allow Gaussian noise to be
added to the likelihoods and then scale the likelihoods by intstd. Of
course this makes noisestd and intstd unidentifiable as logpost can also
be written as logpost ~ N(liks, 1) * noisestd / intstd ** 2. The square may
have helped a bit, but the correlations in the posterior clearly indicated
that the redundancy between noisestd and intstd in this formulation is an
issue.

Consequently, I have now removed noisestd from the model, when
useliks=True. I could have similarly removed intstd, but it felt more
intuitive to assume that the likelihoods should already contain the
necessary noise. Apart from non-decision time this makes the model
deterministic for given likelihoods, meaning that it cannot predict choice
distributions for single trials as long as parameters are fixed. I will
have to see whether this creates any issues.