fix bound and noisestd
Accidentally I only fixed noisestd in the last run and bound was inferred
together with intstd (cf. result 201907122040). The fit was reasonable,
but not as good as with bound fixed and noisestd and intstd inferred. Also,
there is clearly posterior correlation between bound and intstd.

This change now fixes bound and noisestd while only intstd is inferred.