preparing inference with EEG data for all subjects
The last run with fixed noisestd and bound on subject 18 showed great
promise with reasonable fit and without large posterior parameter
correlations. So I now want to run this across all subjects.

EEG data for subjects 1, 9, 10 and 25 is missing. Changes in helpers will
identify the corresponding available subjects in the eegdir.